[PWA] Add client_mode_valid_and_specified in manifest parsing logic
This change updates the manifest parsing logic so that manifests not
having a defined launch handling client mode store additional
information. This is done by modifying the C++ struct parsing logic to
store a std::optional<ClientMode>, so as to reduce the possibility
of malformed data obtained via the manifest mojo.

This is required for navigation capturing to distinguish between apps
that have the launch handler mode specified in the manifest vs apps that
perform the default behavior.

Both of these were done in the same CL due to how the types are re-used
in both places, so changing one caused our parsing tests to fail (and
required creating the launch handler with the extra arg too).

Bug: 385362495
Change-Id: I2c9d8f6933304aedd912fbcd1462f352e05e31a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6168555
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409074}