[mediapipe] Fix handling of com_google_audio_tools in update script
The logic introduced in https://crrev.com/c/6088187 to apply
com_google_audio_tools_fixes.diff was broken, because
`git apply --directory=<directory> <patch>` expects `<directory>` to be
the path relative to the git repository root, not relative to the
current directory.

The original approach would have been correct if mediapipe was in a
separate submodule, but that's not the case.

Bug: 377618534
Change-Id: I378e591602a15701d69d8367fc5362976ef1fe4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179432
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1409075}