[iOS][IPH]Improve the code for testing promo
While working on testing a promo, I found the code slightly complex to
read. Because, a string could mean enabling a single feature, all
feature or no feature dependings on whether it’s empty, a feature name,
or a string that is distinct from feature name.

I introduce an enum to state whether 0, 1 or all features should be
enabled, and a small class that contains this enum and the feature name
if relevant.

I then use this class everywhere relevant in the code.

It should be noted that, while the code is partially in components, it
seems to be used by iOS only.

I also changed variables containing `fet` to `FET`, as FET is an
abbreviation.

Bug: None
Change-Id: I49cc349c08d7c2de30b28dd8ddfda158a25ff443
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6172455
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409079}