PA: new bidder worklet thread selector
The current method of assigning a thread to a generateBid task is to
use a hash of the group's joining origin. This could cause an
imbalance in the loads of the worklet threads. Instead, round-robin
thread assignment for joining origins and enforce a maximum allowed
load imbalance.

This CL implements the new thread selection logic, and a subsequent
CL will use this logic in the bidder worklet.


Bug: 390444489
Change-Id: I0860153154d03cab1f21f84b1af7ea0651d4cd62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179863
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: Abigail Katcoff <abigailkatcoff@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409097}