Allow chrome://histograms selecting histogram while monitoring
In crbug.com/40864008 it was discussed that selecting histogram while in
monitoring mode isn't really supported because as soon as new data comes
the selection gets invalidated. The only way to monitor single histogram
as to select them before getting into monitoring mode.

This attempts to fix that by doing the following:

1. Make `StartMonitoring` ignore the query taking the baseline of ALL
the histograms even if one is preselected. This is done to allow
selection change without changing a base point.

2. Make `GetDiff` not ignore the query returning only the histogram that
is actually requested by a frontend page.

3. Don't force `requestHistograms` when hash selection changes during
monitoring mode. This mixes monitoring data which are diff from baseline
with a global histogram data. With monitoring mode the updated data will
be pulled with the next fetch anyway (within 1s).

4. Link header to `#` allowing to reset selection (aka. `query`) without
refreshing the whole page and therefore the baseline. This allows
changing selection live and opens potential for selecting multiple
histograms.

There is a small risk of increased memory footprint for users doing a
lot of selective monitoring as each of them will now keep the whole
"baseline" set of histograms. However it seems like an extremely obscure
use case and the footprint is not enormous anyway O(kB).

Tested:

 * Updated and added new coverage to automated testing
 * Checked the functionality with local build on gLinux

Change-Id: I8fbf6e7fc138e2639e2d5ea5dff2d48a5666075a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175086
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Andrzej Fiedukowicz <afie@google.com>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409115}