android: Turn off RBE cache for benchmarks
When benchmarking build speed, the RBE cache introduces a source of
nondeterminism that can make the second time the benchmark is run way
faster. This is not deterministic as sometimes subsequent runs do not
benefit from the cache and sometimes it does, perhaps due to resource
constraints.

Bug: 391348954
Change-Id: I28bb0158a57c82855e0eafe091a7322f3b0ba2ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6185557
Reviewed-by: Mohamed Heikal <mheikal@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409116}