[WebAuthn] Add metric to show when passkeys arrive after autofill
We have several metrics around the timing for passkey retrieval and
display, but not one to directly show how often autofill UI shows that
is missing passkeys because they took too long to become available.

This adds a metric to record that. It emits when the passkey list is
returned.

Bug: 367563090

Change-Id: Ie0f1ad69dfbfc058afb99c78ef096acf63c3349d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6180106
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409128}