Fix crash in recording OCR language when there are no results.
When OCR returns no results, ScreenAI service tried to record the most
detected language in the non existing result and crashed.
A TODO is added to extend testing to cover this situaiton, but the test
is not added in this CL to make it easier to merge the fix in previous
branch.

AX-Relnotes: n/a
Bug: 391040508
Change-Id: I06a3ec72921b486056a4fd981ef876efbf954afb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6182128
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Commit-Queue: Kyungjun Lee <kyungjunlee@google.com>
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1409132}