[Extensions] Remove system indicator API code
The chrome.systemIndicator API was used for handling the system tray
icon for the Hangouts extension and at this point is only allowlisted to
that extension (along with the development versions and associated
tests).

Since that extension is long gone at this point we can remove the API
and any related code entirely.

Bug: 389978739, 40257657
Change-Id: I85c9f2f18b77944c894dc3c8a9d1239d98b7da9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6173200
Commit-Queue: Tim <tjudkins@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409165}