[lensoverlay] Fix overlay sometimes covering dev tools
In http://shortn/_OfpJw3GrSu, I moved the overlay view to be a sibling
of the contents web view, instead of a child. Due to that change, the
coordinate system being used to set the overlay bounds was local to the
contents web view, not to the new parent. Since the contents web view
and overlay web view are siblings, they are in the same coordinate
system, so this CL changes our overlay view to always have the same
bounds and position as the contents web view.

Bug: 391260903
Change-Id: I9cb8ed96a7d57f08ee346e522d9d7172a0a924e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187293
Reviewed-by: Juan Mojica <juanmojica@google.com>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1409167}