[ios] Add profile pref observing to ReminderNotificationClient
This change adds profile pref observing to ReminderNotificationClient
so that it can schedule notifications when a new reminder is added to a
dictionary pref. In order to do this, it also needs to observe the
ProfileManagerIOS in order to observe the prefs for each loaded profile.

Fixed: 389911417
Change-Id: I40856bcaf3254e4a49d2307a031f3c5730b0554b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6180950
Commit-Queue: Huiting Yu <huitingyu@google.com>
Reviewed-by: Huiting Yu <huitingyu@google.com>
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1409178}