bluetooth: Expose paired BT devices in Android behind a feature flag
Note we only start pulling in devices when a caller calls
BluetoothAdapter::GetDevices() or starts a discovery session so that we
we give the frontend a chance to request for Bluetooth permissions.

Unlike implementstions on some other platforms, the Android
implementation doesn't create separate classes to differentiate classic
and low energy devices. This is because Android uses BluetoothDevice to
represent both types of devices, whose equality solely depends on the
addresses of remote devices. If we choose to separate implementations of
classic and low energy devices, we may encounter issues on devices of
dual transports as both implementations may interact with the same
instance of BluetoothDevice. This is similar to BlueZ and Floss,
implementations on Linux.

Bug: 380128287
Test: Paired devices are listed.
Test: device_unittests on Linux and Android.
Change-Id: I2299277153758ececb188fbd44d1aa696b2f645d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6114763
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Reviewed-by: Jack Hsieh <chengweih@chromium.org>
Commit-Queue: Garfield Tan <xutan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409206}