[passage-embeddings] Moves PassageKind to passage_embeddings namespace
Also renames the enum to `PassagePriority` for consistency with the
mojom enum in case the two type definitions were to be combined in the
future.

This CL does not do that yet because the embedder cares about user
initiated and passive passage priorities while history embeddings
search has a third lower priority type for rebuilding the DB which
is treated the same as passive priority from the embedder perspective.

Bug: 390241271
Change-Id: I57569e310faea6a0d3ed048faf1f38f4bb493de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6185554
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Orin Jaworski <orinj@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1409222}