[User Education] Carry ComputedData across precondition lists
Preconditions in the "wait for" list should have access to computed
data from the "required" list.

This allows for the lifecycle data to be used by the anchor precondition
to find the correct anchor for rotating promos.

Bug: 369887098
Change-Id: Ic620b9308450d0b26965e4216dfff59efc6d7835
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6181416
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409230}