[FedCM] Ensure early return when onDismiss is invoked on Android
When show* methods are invoked, sometimes it ends up with onDismiss,
which causes the backend to clear the majority of its state. This CL
plumbs this information from the UI code back to the caller, similar
to how it is already done on desktop. This ensures that a caller will
early return when onDismiss is invoked. In particular, this tentatively
fixes a CHECK being hit due to the members being cleared when onDismiss
is invoked.

Bug: 390790111
Change-Id: I2dbedc130b68d23d246754afc6e3107c5faca8ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6188080
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409324}