[lensoverlay] Remove percent decoding when getting text fragments.
This could cause text fragments with percent encodings to fail to parse
if they included any characters that are used to find the prefix and
suffix of the fragment such as `,` and `-`. To fix this, remove any use
of the URL and URLSearchParams interface and extract text fragments from
the `url` string directly.

Change-Id: I4713d414c73f9b1a8e0d520ead8b6e262cfe4b2c
Bug: 383575917
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179629
Reviewed-by: Andy Phan <andyphan@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1409342}