[Ash] Delete unneeded crosapi_ash()->magic_boost_controller_ash() usage
Right now there are two ways to summon the shared instance of
MagicBoostControllerAsh to call its ShowDisclaimerUi() method. One is
through calling the shared instance of CrosapiAsh first, which itself
has a reference to a shared MagicBoostControllerAsh instance. The other
way is to call ash::MagicBoostControllerAsh::Get() to reach the same
shared instance.

This CL deletes all usage of the former way of accessing this shared
instance. And as an intended byproduct, this CL breaks the unnecessary
circular dependencies of several feature code on //c/b/ash/crosapi.

From a bigger perspective, this CL further weakens the status of
//c/b/ash/crosapi, which is subject to sunsetting and deletion, with the
sunsetting of Lacros.

Bug: 354710097
Test: tryjob
Change-Id: I2fbf3c0cb31cf7f71bd09e95e30ee0235e485a56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6185179
Reviewed-by: Georg Neis <neis@chromium.org>
Commit-Queue: Di Wu <diwux@google.com>
Cr-Commit-Position: refs/heads/main@{#1409349}