Prevent send tab to self from trying to show for web apps.
This is a speculative fix for the linked bug. I believe the crash is
possible if the last active browser is a web app when a tab is sent to
self from another device and causes the crash in
SendTabToSelfToolbarIconController because the
PinnedToolbarActionsContainer will not exist. This changes the behavior
to store the send-tab-to-self entry if the browser is not a normal type
and adds a regression test.

Low-Coverage-Reason: OTHER uncovered lines is existing untested functionality and planned to soon be removed with the launch of Toolbar Pinning.
Bug: 391064428
Change-Id: Ib238b662d4bfa44ff98134ee2dbf61d46ce773cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179189
Reviewed-by: David Pennington <dpenning@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409406}