Use mojo_base.mojom.UnsafeSharedMemoryRegion in protected_buffer_manager.
handle<shared_buffer> was deprecated, and so blocked by PRESUBMIT
on touching the API, including file move.

This CL gets rid of it. This is API breaking change of Stable
interface, but the counter part of this API in CrOS is not using
this API yet, so safe to change.

BUG=386148857
TEST=Tryjob

No-Presubmit: True
Change-Id: Ibe19cfd1b6f0f47aa53ecbcb50208c31b0164cae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6171482
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409441}