Fix to show favicons for tabs in a saved tab groups in app menu
Favicons are now reliably shown for tabs in saved tab groups from the
app menu. Caching favicons in-memory isn't scalable if we have a lot of
saved tab groups even those that aren't loaded. Instead we load it
on-demand from the favicon service when a user is using the re-call UI.
Due to this removed the code to cache favicons from the web contents
listener since it results in only caching favicons for specific
scenarios. Note this doesn't address the issue from context menu in the
bookmark bar. That fix will happen in a follow-up CL.

Bug: 343279726
Change-Id: I737369e7ac24081f600b1e02a0e6b7007f567abc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6186471
Commit-Queue: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1409469}