[Code Health Rotation] Clean up URLCanonTest
As part of the code health rotation, we should remove the stale feature
flag `StandardCompliantNonSpecialSchemeURLParsing`. This feature was
launched in M130. See https://crbug.com/388143055 for details.

This CL just did the minimum change required to remove the feature flag.
Further cleanup could be possible, but that can be done in another CL.

Bug: 388143055
Change-Id: I2b2a4d6cd4505fa6a968419923137e47b39c45cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6185016
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409491}