Fix empty accessibility tooltip
SetText() that was previously used to set the text near the icon. It
turns out that it sets the accessibility tooltip too. So in some cases
the accessibility tooltip could be an empty string, which is not desired
and may lead to a crash.

Bug: 381054335
Change-Id: Ia1d5a328443a26415457e3ff63b0f16ef346f498
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6182363
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1409526}