[iOS][MIM2] Create AuthenticationFlowInCurrentProfile class
This class is in charge to finish the sign-in flow after profile
switching happened.
This class is triggered by AuthenticationFlow, to continue the sign-in
flow with the newly loaded profile.

All the steps from AuthenticationFlowInCurrentProfile are based from
the AuthenticationFlow steps.

In a later patch, AuthenticationFlow will be able to use
AuthenticationFlowInCurrentProfile even when there is no profile
switching.

Unittests will be added in a following patch.

Bug: 375605482
Change-Id: Ic5cbfc2c350d23351efa9543d41eecda16a929ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6182151
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409538}