kiosk: Launch pre-installed Chrome app when update fails
Currently Kiosk tries to install/update the Kiosk Chrome app and its
secondary apps at every launch, and fails launch if the install fails.
This behavior makes sense for the very first install, but we can be more
lenient during updates.

With this change Kiosk will try to proceed with the installed Chrome app
that already exists in the profile, when an update from the cache fails
during launch.

Bug: 383975759, 389635579
Test: tast run <DUT> 'kiosk.Smoke.*'
Test: testing/xvfb.py tools/autotest.py -C out/Default chrome/browser/ash/app_mode/test/minimum_chrome_version_browsertest.cc
Change-Id: I4e3508dff9eb90620476230331b8fc7a50f3992c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6097471
Reviewed-by: Polina Bondarenko <pbond@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Edman Anjos <edman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409543}