Extend unit test coverage for bookmark writer
Doing this ahead of refactoring and behavior changes in child CLs as the
current coverage is quite light.

* Add change detector tests that directly validate the contents of the
output file. The format is intended to be stable, and supported in other
browsers - so testing that Chromium can import it successfully may not
be sufficient.
* Add coverage for a couple of additional test cases: no bookmarks, and
bookmarks in the bookmark bar only.

Bug: 390643468
Change-Id: I370ad80fd0638aed52760092456757d6771b4653
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6178284
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Auto-Submit: James Lee <ljjlee@google.com>
Commit-Queue: James Lee <ljjlee@google.com>
Cr-Commit-Position: refs/heads/main@{#1409574}