[DBSC] Do not terminate a session when required cookies are not received
A small number of DBSC rotation requests per day that fail with
kServerUnexpectedResponse, which leads to a session termination.

We believe that this error has either a transient nature or caused by a
bug on the client side (for example, a Mojo cookie observer API might
lose a connection).

To drive the number of wrongly terminated sessions down, this CL changes
kServerUnexpectedResponse error status from a persistent error to a
transient error. The debug header is update with the information about
which cookies were missing exactly.

Fixed: 391166740
Change-Id: I46c77769e8e5c098d3f160b3d07f921b4feffdc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187116
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1409624}