Fix tools/translation/upload_screenshots.py for Cog.
Git isn't available in Cog, and so when searching for grd files, we can't do a `git ls-files -- *.grd`.  Previously, we tried walking the repo to find all instances of _.grd_ files.  However, that meant that we were listing _.grd_ files in submodules, which we don't want.

Instead, we'll just skip the checks which involve _.grd_ file paths which are added to the repo but not added to the expectations file.  Since presubmits run in CQ, this shouldn't result in invalid repo states.  It's just a bit inconvenient in that edge case.

Alternatives considered:

- Use the grds present in the expectations files.
  - This is possible, but is equivalent to the current implementation.  (Effectively, we'd be saying that the expectation file is equivalent to the expectation file.)
- Make an RPC call to simulate the git ls-files result.
  - This would require getting the correct credentials, and would make the implementation significantly more complex.

For unit tests, I've added a new test data file for each test case involving Cog.  While we could have each of the missing files in one test data file, doing so would enforce the order errors are added to the list of errors.  So having a separate check for each is less prone to toil if the implementation is changed.

Change-Id: I5552185b817eb4bf962b23d2d6fad15ad9ddc020
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6177203
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Terrence Reilly <treilly@google.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#1409647}